Use existing GitHub usage infrastructure to run an audit of alphagov membership.
The same lambda is re-used. Because doing the full audit would
take time the process is broken down into multiple invocations
using SNS to pass from one to the next.

The start of the audit creates 3 sns messages to get
- org members
- org teams
- org repos

Then for each team and repo it publishes SNS messages to invoke
the lambda to audit team members and repo access by teams and
contributors.