Add an id UUID to every event in the audit process.
This enables us to correlate all the events from a single audit run.
Increase the lambda concurrency from 4-6. We should be currently
running at around 1/2 the github rate-limits but I want to leave
some headroom.
Make cron schedule optional in terraform so we can leave non-prod
unscheduled.
Add event input to the cron triggers so that the lambda is invoked
with the correct action.